+++ Modules/Microsoft365DSC/DSCResources/MSFT_AADGroupsNamingPolicy/MSFT_AADGroupsNamingPolicy.psm1
@@ -354,8 +354,11 @@ function Export-TargetResource
             $dscContent += $currentDSCBlock
         }
 
-        Save-M365DSCPartialExport -Content $currentDSCBlock `
-            -FileName $Global:PartialExportFileName
+        if ($currentDSCBlock)
+        {
+            Save-M365DSCPartialExport -Content $currentDSCBlock `
+                -FileName $Global:PartialExportFileName
+        }
 
         Write-Host $Global:M365DSCEmojiGreenCheckMark
         return $dscContent